Update to the remaining anticipated 0xc features.
Changes:
 - Convert to block signatures.
 - Convert to section opcodes.
 - Minimally update Release/AngryBots.js to work with the new JS API.
 - Include a binary-encoded wasm file in 0xc format.

I made up a syntax for block signatures in the flat text format, taking
care to allow for adding arguments and multiple results in the future.
For example:

```
      block $label28 () : (i32)
```

Is a `block` no arguments and one return type which is i32.

However, I'm not attached to this particular syntax, and it can easily
be converted to something else if desired.