Remove demo
This was convenient when we first launched, but now there are many demos
on the web of WebAssembly, so we don't really need to host one here.

In addition, there seem to be issues w/ the Tanks demo
(see #87, #88, #110, #126), and no one is maintaining it. Maybe in the
future we'll want to link to external demos, but for now let's just
remove.