Update features for wasmtime 15 (#358)
https://github.com/bytecodealliance/wasmtime/releases/tag/v15.0.0

> The WebAssembly threads, multi-memory, and relaxed-simd proposals are
now enabled by default.
> https://github.com/bytecodealliance/wasmtime/pull/7285

I wasn't sure whether to specify the version as "15" or "15.0" (or even
"15.0.0"). It seems to be inconsistent in the table. I used 15, as
wasmtime's releases are similar to a browser where on a fixed schedule
they bump the major version and bugfixes are done as patch versions
(minor versions aren't used).