Rename roadmap to "feature extensions" (#361)
When this page was created along with the wasm MVP, all the listed
features were in the future, so it served as a roadmap for what was to
come. Now it includes all the feature extensions that have been added
since, and its primary usefulness is actually for developers to look up
which engine versions implement which features (including past and
near-future features which have implementations behind flags). So a
better name than "roadmap" or "future features" would be something like
"feature extensions".

Also fixes #329