Update Node status for GC (#369)
As of Node.js v22.0.0-nightly20240104084d761dfc (Jan 5, 2024), node is updated to
v8 11.9.169.7, which supports GC by default with the final binary encoding (which is not
compatible with previous unreleased versions). So any version of node v22 and above
built after this date should work with no flags, and any previous version will not work
with current tools, even with a flag.