Track and update all the implemented features (#390)
After seeing https://labs.leaningtech.com/blog/branch-hinting yesterday,
I was quite surprised to read that not only is this proposal done, but
it's also implemented in all the browsers, and yet it's not even tracked
in the `features.json` file.

So I took it upon myself to make sure that all features that are in at
least phase 2  are tracked and up to date. 

With there being a lot more proposals that are being tracked now, I
thought it would also make sense to properly reflect the phases on the
website, instead of splitting them into just two sections.