Add GraalWasm to "Feature Extensions" page. (#396)
(Disclaimer: I'm on the [GraalVM](https://www.graalvm.org/) team that
maintains GraalWasm).

This PR adds GraalWasm to the [Feature Extensions
page](https://webassembly.org/features/). GraalWasm is an
[open-source](https://github.com/oracle/graal/tree/master/wasm)
WebAssembly runtime written in Java using our [Truffle
framework](https://www.graalvm.org/latest/graalvm-as-a-platform/language-implementation-framework/).
We introduced [it in
2019](https://medium.com/graalvm/announcing-graalwasm-a-webassembly-engine-in-graalvm-25cd0400a7f2)
and plan to declare it as stable in the upcoming 24.1.0 release on
September 17, 2024.

Note that https://graalvm.org/webassembly will become available in the
next couple of days, we are still working on it.

I thought I open this PR now because it seems that the Feature
Extensions page may need a little tweaking. With the additional
GraalWasm entry, the table starts to become wider than the content box,
causing an awkward horizontal scrollbar (see below). I'm happy to
incorporate any fixes or improvements for this.


![image](https://github.com/user-attachments/assets/fd148227-6d6c-4486-94b4-af8c8217881c)