Add Chicory to 'Feature Extensions' page. (#433)
Hi there :wave: thanks for maintaining the WebAssembly official site!

This PR adds Chicory to the [Feature Extensions
page](https://webassembly.org/features/). Chicory is an
[open-source](https://github.com/dylibso/chicory) pure java WebAssembly
runtime written, it includes an interpreter and an experimental compiler
from wasm to java bytecode.
We announced [it in
2023](https://www.javaadvent.com/2023/12/chicory-wasm-jvm.html) and we
keep evolving and adding new features and proposal implementations.

I'm a maintainer of the project, happy to answer any question!